+++ react/apps/grid/tailwind.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  content: [
+    './src/**/*.{js,jsx}',
+    '../../packages/grid/src/**/*.{js,jsx}',
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};